Provide OpenGL drivers on non-NixOS
To function openauto requires OpenGL, specifically EGL on wayland.
This is problematic with Nix as OpenGL drivers are managed in a impure
way under NixOS, with them existing at /run/opengl-driver.  That's not
the case with Raspberry Pi OS (Debian) as they exist at
/usr/lib/aarch64-linux-gnu, along with all the other system libraries.

There's no easy way to get the system libraries to work on Debian.  The
best course of action would be to copy a few libmesa libraries into
~/.opengl-driver, patchelf their *rpath* to the system path, and point
LIBGL_DRIVERRS_PATH to /usr/lib/.../dri.  But, we run into an issue
where libLLVM as compiled by upstream has an RPATH of ${origin}/../lib,
and that messes everything up as that turn /usr/lib/aarch64-linux-gnu
into /usr/lib/lib, and child libraries then fail to load.

"nixGL" is a tool that will let you run commands from nix that require
OpenGL on non-NixOS systems, but it's a runtime evaluation, which isn't
compatible with a car environment of
(Offline || Metered Networking) && Instant On.

Nvidia GPU's won't work out of the box, and that's probably not what
people use in a car, so with Nvidia use base dash package and run it
with nixGL.